UIIN-3175: Revert changes
